Add basic lttng instrumentation
Added first lttng tracepoints definition file and changed makefiles to
compile and link the right libraries. Currently it's not optional, just
a hack. Will needed autoconf detection or should be made optional.
Hacked in a couple of tracepoints in erl_process.c around the call to
check for I/O.